image_processing/pyproject.toml
[project]
name = "image-processing"
version = "0.1.0"
description = ""
authors = [
    {name = "Matas Seimys",email = "matas.seimys@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"

[tool.poe]
env = { PYTHONPATH = "." }

[tool.poe.tasks]
test = "pytest -v"

[tool.poe.tasks.processor]
cmd = "python image_processing/workers/image_processor.py"

[tool.poe.tasks.run]
use_exec = true  # or we get stuck inside asyncio loop
cmd = "uvicorn image_processing.api.main:app --host 0.0.0.0 --port 5000 --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
sqlmodel = "^0.0.24"
alembic = "^1.15.1"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
confluent-kafka = "^2.8.2"
pillow = "^11.1.0"
open-clip-torch = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
poethepoet = "^0.33.0"

[tool.ruff]
line-length = 130
indent-width = 4

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "image_processing",
]
